requirements.txt
Flask
Flask-Login
fpdf2
pandas
openpyxl